telegram_bot_service/requirements.txt
pyjwt==2.10.1
requests==2.32.5
python-telegram-bot==22.4
regex==2025.9.1
python-dotenv==1.1.1
ftfy==6.3.1
yandex_cloud_ml_sdk==0.15.0